Fix: Documentation is blank for hints with in hint sets
This commit also contain a refactor for how we generate the
documentation in the website.

Fix #617
