+++ helpers/update-site/common.js
@@ -1,0 +1,22 @@
+const fs = require('fs');
+const path = require('path');
+
+const mkdirp = require('mkdirp');
+
+const setShellJSDefaultConfig = (shelljs) => {
+    shelljs.config.fatal = true;
+};
+
+// Create folder if not exist before writing file.
+// Reference: https://stackoverflow.com/a/16317628
+const safeWriteFile = (dir, content) => {
+    const folderPath = path.dirname(dir);
+
+    mkdirp.sync(folderPath);
+    fs.writeFileSync(dir, content); // eslint-disable-line no-sync
+};
+
+module.exports = {
+    safeWriteFile,
+    setShellJSDefaultConfig
+};
