+++ src/content/404.md
@@ -1,5 +1,5 @@
 ---
-category: "placeholder-pages"
+section: "placeholder-pages"
 layout: "error"
 originalFile: "404.md"
 permalink: "404/index.html"
