+++ src/webhint-theme/helper/utils.js
@@ -7,48 +7,38 @@ const isIndexPage = (page) => {
 };
 
 const isGuideIndexPage = (page) => {
-    return isIndexPage(page) && page.title.toLowerCase().replace(' ', '-') === page.category;
+    return isIndexPage(page) && page.title.toLowerCase().replace(' ', '-') === page.section;
 };
 
 const isSectionIndexPage = (page) => {
-    return isIndexPage(page) && page.title.toLowerCase().replace(' ', '-') !== page.category;
+    return isIndexPage(page) && page.title.toLowerCase().replace(' ', '-') !== page.section;
 };
 
-const sortPageByAlpha = (l, r) => {
-    if (l.title > r.title) {
-        return 1;
-    }
-
-    return -1;
+const isGoodToToC = (page) => {
+    return !isGuideIndexPage(page) && page.tocTitle && page.contentType !== 'hints-category' && page.isMultiHints !== 'true';
 };
 
-const sortPageByIndex = (l, r) => {
-    const indexL = l.index;
-    const indexR = r.index;
-
-    if (isSectionIndexPage(l)) {
+/**
+ * Sort pages alphabetically given a `title` unless the file is `summary.md`
+ * in which case it will be the first
+ * one.
+ * E.g.:
+ * Server configurations:
+ *  - Examples of server configurations (`summary.md`)
+ *  - Basic server configuration for Apache (`apache.md`)
+ *  - Basic `web.config` for IIS (`iis.md`)
+ */
+const sortPages = (l, r) => {
+    if (l.originalFile.endsWith('/summary.md')) {
         return -1;
     }
 
-    if (isSectionIndexPage(r)) {
+    if (r.originalFile.endsWith(`/summary.md`)) {
         return 1;
     }
 
-    // Check if one of the pages is the index page before proceeding to index comparison
-    // If index page checking is combined with index comparison, the following edge case will return -1 instead of 1:
-    // - indexL:01
-    // - indexR: undefined; isSectionIndexPage(r) = true
-
-    if ((indexL && !indexR) || indexL < indexR) {
-        return -1;
-    }
-
-    if ((!indexL && indexR) || indexL > indexR) {
+    if (l.title.toLowerCase() > r.title.toLowerCase()) {
         return 1;
-    }
-
-    if ((!indexL && !indexR) || (indexL === indexR)) {
-        return sortPageByAlpha(l, r);
     }
 
     return -1;
@@ -77,29 +67,9 @@ module.exports = {
             return !hint.isSummary;
         }).length;
     },
-    /**
-     * Returns all the pages under a given `title` sorted alphabetically
-     * unless the file is `summary.md` in which case it will be the first
-     * one.
-     * E.g.:
-     * Server configurations:
-     *  - Examples of server configurations (`summary.md`)
-     *  - Basic server configuration for Apache (`apache.md`)
-     *  - Basic `web.config` for IIS (`iis.md`)
-     */
-    getPagesByToCTitle: (title, pages) => {
-        return pages[title].filter((page) => {
-            return page.contentType === 'details';
-        }).sort((a, b) => {
-            if (a.originalFile.endsWith('summary.md')) {
-                return -1;
-            }
-
-            if (b.originalFile.endsWith(`summary.md`)) {
-                return 1;
-            }
-
-            return a.title > b.title;
+    getPagesBySection: (pages, section) => {
+        return pages.filter((page) => {
+            return page.section === section;
         });
     },
     getSignalIssueQuery: (root, title, directory) => {
@@ -108,20 +78,17 @@ module.exports = {
 
         return `title=${encodeURIComponent(issueTitle)}&body=${encodeURIComponent(issueContent)}`;
     },
-    getSortedToCTitles: (pages) => {
-        return Object.keys(pages).sort();
-    },
     /**
      * Return the `href` value to link to the target page from the current one.
      * for example when the current page is `hint-amp-validator`:
-     * page = { category: 'user-guide', permalink: 'docs/user-guide/hints/hint-amp-validator/index.html' }
+     * page = { section: 'user-guide', permalink: 'docs/user-guide/hints/hint-amp-validator/index.html' }
      * getTargetPagePath(page, 'docs') => '../../../'
-     * getTargetPagePath(page, 'category') => '../../'
+     * getTargetPagePath(page, 'section') => '../../'
      */
     getTargetPagePath: (page, target) => {
-        const { category, permalink } = page;
+        const { section, permalink } = page;
         const parts = permalink.split('/');
-        const targetPage = target === 'category' ? category : target;
+        const targetPage = target === 'section' ? section : target;
         let count = parts.length - parts.indexOf(targetPage) - 2;
         let finalPath = '';
 
@@ -131,6 +98,35 @@ module.exports = {
         }
 
         return finalPath;
+    },
+    /**
+     * Get the ToC items from a list of pages and sort the elements.
+     */
+    getToCItems: (pages) => {
+        const items = pages.reduce((acc, page) => {
+            const tocSectionTitle = page.tocTitle;
+
+            if (tocSectionTitle && !acc[tocSectionTitle]) {
+                acc[tocSectionTitle] = [];
+            }
+
+            // non-guide-index pages and non-multi-hint pages
+            if (isGoodToToC(page)) {
+                acc[tocSectionTitle].push(page);
+            }
+
+            return acc;
+        }, {});
+
+        const sortedKeys = Object.keys(items).sort();
+        const result = sortedKeys.map((key) => {
+            return {
+                pages: items[key].sort(sortPages),
+                title: key
+            };
+        });
+
+        return result;
     },
     hasSubPage: function (id) { // eslint-disable-line object-shorthand
         return (id === 'docs' || id === 'about');
@@ -154,33 +150,6 @@ module.exports = {
         // If the markdown Content should be used.
         const guideIndexes = ['contributor guide', 'user guide'];
 
-        return page.contentType === 'details' || guideIndexes.includes(page.title.toLowerCase()) || (page.contentType === 'index' && page.tocTitle !== 'hints');
-    },
-    // Sort out `Developer guide` or `User guide` pages
-    sortPagesByCategory: (allPages, category) => {
-        const pages = allPages.reduce((acc, page) => {
-            if (page.category === category) {
-                const tocSectionTitle = page.tocTitle;
-
-                if (tocSectionTitle && !acc[tocSectionTitle]) {
-                    acc[tocSectionTitle] = [];
-                }
-
-                if (!isGuideIndexPage(page) && page.tocTitle) { // non-guide-index pages
-                    acc[tocSectionTitle].push(page);
-                }
-            }
-
-            return acc;
-        }, {});
-
-        // Sort pages numerically first, and then alphabetically
-        for (const section in pages) {
-            if (pages.hasOwnProperty(section)) {
-                pages[section] = pages[section].sort(sortPageByIndex);
-            }
-        }
-
-        return pages;
+        return page.contentType === 'details' || guideIndexes.includes(page.title.toLowerCase());
     }
 };
