+++ src/webhint-theme/layout/partials/hints-category.ejs
