+++ src/webhint-theme/layout/partials/hints-index.ejs
